Add Notification Banner block
Because the <p> tags within the banner require their own specific class, I've opted to make that a textarea rather than a wysiwyg for now.

That means that they'll only be plain text (unless you add the HTML for a link), but I think that's ok for now. We can always revisit if needed.

Resolves: https://trello.com/c/7qyeExDS/23-add-notification-banner-to-plugin