Allow admins to toggle blocks on or off across the site
Before: All blocks in the plugin were enabled by default.

Now: Admin users can toggle each block on or off under Settings > GOV.UK Components

In order to enable this, there's been a bit of refactoring. The blocks all implement an iBlock interface, ensuring they have an `init()` method (activating the block), and an option name and a display name. All the blocks are passed to a BlockController class. This is responsible for:
a) Generating the list of blocks that are available on the options page
b) Generating the list of blocks activated by default (currently all of them)
c) Activating the blocks the admin has chosen to make active.

The BlockController class is passed to Options. This is responsible for:
a) Generating the options page & the fields on it (using the block options supplied by the BlockController)
b) Telling the BlockController which blocks to activate, based on the admin's selections

This should make it easy to add more blocks in future, as all we'll have to do is add a new class that implements iBlock, new it up and pass it into the BlockController. The corresponding options etc. will be added automatically.

Resolves: https://trello.com/c/qrc3tq7J/30-plugin-should-allow-admins-to-toggle-component-blocks