Added new block Category 'Custom'
Before: we should give a name to our component blocks so they can be easily identified; it can be gov.uk or design-system (or something like it) for now, eventually, it would be better to not refer to gov.uk or gds. Ideally, we would use the site or client name (or an abbreviated version of it) to indicate that these were their blocks.

After: we decided to create a new block category with the slug 'govuk-custom' and the name 'Custom' on top of the other categories, so the component blocks are now visible as a group and easily identified. Ideally in the future we should add an option into the plugin to change the category label with any string: the site name, the client name, etc.

Resolves: https://trello.com/c/JFYxV5ZO/31-name-our-component-blocks-so-they-can-be-identified