Add option for editable IDs on accordion section headings
Some clients have requested the ability to attach IDs to headings in
accordions, so that they can then link directly to that section of the
accordion.

This commit adds that option. To keep it unintrusive for those who don't
wish to use it, it requires a checkbox to display the ID field. The ID
will only be rendered in the front-end if the checkbox is ticked and
the ID field is not empty.

For: https://trello.com/c/aT1vY4Pf/8-iterate-on-or-rebuild-accordion-block