Add callback to display GovUK phase banner
Note that we're using a Flatpack specific hook here,
which makes it more difficult for non-dxw code to use
this plugin.

For now, this is a pragmatic choice because we need a
hook that exists in the parent theme, but it might be
wise to add a hook for wp_body_open or similar.