Phase banner is applied with wp_body_open
Previously we added the phase banner to pages in response
to a dxw specific hook that we provide in a parent theme.

This makes the banner unusable for anyone outside dxw,
but using a standard WordPress hook would not work on
our sites.

This commit adds the phase banner in response to the
wp_body_open hook, which is provided by WordPress Core.
This means that non-dxw themes can use the component,
but any theme using both hooks will presumably find the
banner appears twice.

See:
https://developer.wordpress.org/reference/hooks/wp_body_open/